% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BergerParkerP.R
\name{BergerParkerP}
\alias{BergerParkerP}
\title{Parallelised Berger-Parker's diversity index}
\usage{
BergerParkerP(rasterm, w, na.tolerance, debugging)
}
\arguments{
\item{rasterm}{Input raster data as a matrix or suitable object convertible to a matrix, containing the values on which to compute the index.}

\item{w}{Integer, representing half of the side length of the square moving window used in calculations.}

\item{na.tolerance}{Numeric value between 0.0 and 1.0 indicating the proportion of NA values tolerated in each moving window over the input rasterm. If the proportion of NA's exceeds na.tolerance, the result for that window is set as NA. The default is 0.0, indicating no tolerance for NA values.}

\item{debugging}{Logical, when TRUE, additional console messages are printed for debugging purposes. Defaults to FALSE.}
}
\value{
A matrix or a list of matrices with the Berger-Parker index computed through a moving window of the specified size.
}
\description{
This function computes the Berger-Parker diversity index in a parallelised manner, 
improving performance on suitable hardware for large datasets.
}
\details{
Parallelised Berger-Parker's diversity index
}
\examples{
\dontrun{
# Generate a sample matrix
data_matrix <- matrix(sample(c(NA, 1:10), 100, replace = TRUE), 10, 10)

# Compute the Berger-Parker index with parallel processing
result <- BergerParkerP(rasterm = data_matrix, w = 1, na.tolerance = 0.1, debugging = TRUE)
}

}
\seealso{
\code{\link{BergerParker}} for the non-parallel version of the Berger-Parker index calculation.
}
\author{
Marcantonio Matteo \email{marcantoniomatteo@gmail.com}

Martina Iannacito \email{martina.iannacito@inria.fr}

Duccio Rocchini \email{duccio.rocchini@unibo.it}
}
\keyword{internal}
