\docType{data}
\name{copNDVI}
\alias{copndvi}
\title{Copernicus Long Term (1999-2017) NDVI Overview (5km)}
\format{RasterLayer:
  \describe{
    \item{NDVI}{Normalised Difference Vegetation Index value (0-255) for each 5 km pixel.}
}}
\usage{
copNDVI
}
\description{
A RasterLayer (EPSG: 4326) of the global average NDVI value per pixel over the period 1999-2017.
}
\references{
\url{https://land.copernicus.eu/global/products/ndvi}
}
\keyword{datasets}
\keyword{NDVI}
\keyword{RasterLayer}