% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.CRE_}
\alias{.CRE_}
\title{Cumulative Residual Entropy}
\usage{
.CRE_(B, base = exp(1))
}
\arguments{
\item{B}{A numeric vector or matrix representing the values for which CRE is to be calculated.}

\item{base}{The base of the logarithm used in the calculation. The default is the natural logarithm (e).}
}
\value{
A numeric value representing the CRE.
}
\description{
This function calculates the Cumulative Residual Entropy (CRE) for a given set of values.
}
\examples{
B <- c(1, 2, 3, 4)
.CRE_(B)
}
