% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.CumRes}
\alias{.CumRes}
\title{Calculate Cumulative Residual Probability}
\usage{
.CumRes(a)
}
\arguments{
\item{a}{A numeric vector or matrix representing probabilities.}
}
\value{
A numeric vector or matrix of cumulative residual probabilities.
}
\description{
This function computes the cumulative residual probability for a given set of probabilities.
}
\examples{
a <- data.frame(V1= c(0.2, 0.3, 0.5), V2 =c(0.2, 0.3, 0.5))
.CumRes(a)
}
