% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.Deltas}
\alias{.Deltas}
\title{Calculate Differences Among Values}
\usage{
.Deltas(P, first = 0)
}
\arguments{
\item{P}{A numeric vector or matrix representing probabilities.}

\item{first}{The starting value for difference calculation.}
}
\value{
A vector or matrix of differences.
}
\description{
This function computes the differences among values of a table, used in probability calculations.
}
\examples{
P <- c(0.2, 0.3, 0.5)
.Deltas(P)
}
