% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.Prob}
\alias{.Prob}
\title{Calculate Point Probability}
\usage{
.Prob(C)
}
\arguments{
\item{C}{A numeric vector or matrix.}
}
\value{
A vector of probabilities corresponding to each point in `C`.
}
\description{
This function computes the probability of each point in a given vector or matrix.
}
\examples{
C <- c(1, 1, 2, 2, 3)
.Prob(C)
}
