% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{.Reorder}
\alias{.Reorder}
\title{Additional supporting functions like `.Reorder`, `.Cumsum`, `.Rev`}
\usage{
.Reorder(a, ax)
}
\arguments{
\item{a, }{ax Additional parameters specific to each function.}

\item{ax}{Additional parameters specific to each function.}
}
\value{
Output varies depending on the function.
}
\description{
These functions provide utility operations like reordering dimensions, 
computing cumulative sums, and reversing order along a specific dimension.
}
