% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_copNDVI.R
\name{load_copNDVI}
\alias{load_copNDVI}
\title{Load Copernicus NDVI data}
\usage{
load_copNDVI()
}
\value{
A `SpatRaster` object representing the Copernicus NDVI data.
}
\description{
This function loads and returns the Copernicus Long Term (1999-2017) NDVI Overview stored within the package.
}
\examples{
copNDVI <- load_copNDVI()
}
