% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenyiS.R
\name{RenyiS}
\alias{RenyiS}
\title{Sequential Renyi's diversity index}
\usage{
RenyiS(rasterm, w, alpha, base, na.tolerance, debugging)
}
\arguments{
\item{rasterm}{Input data, expected to be in a specific format suitable for the calculations.}

\item{w}{Half of the side of the square moving window used in the computation.}

\item{alpha}{The alpha value for the order of diversity in Renyi's Index calculations.}

\item{base}{The base of the logarithm used in the diversity index calculations.}

\item{na.tolerance}{A numeric value between 0.0 and 1.0 indicating the proportion of NA values 
that are acceptable in each moving window over \code{rasterm} during the calculation of Rao's index.
If the proportion of NAs in a window exceeds this value, the result for that window is set as NA.
Otherwise, the index is calculated using the non-NA values. The default is 0.0, indicating no tolerance for NAs.}

\item{debugging}{Boolean; controls the generation of additional, diagnostic messages. 
If TRUE, extra messages are printed primarily for debugging purposes. Default is FALSE.}
}
\value{
A matrix or a list of matrices, each containing the Renyi index values computed using a moving 
window of the specified size.
}
\description{
This function calculates Renyi's diversity index sequentially.
}
\seealso{
\code{\link{Renyi}} for the related function in the context of diversity index calculations.
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, 
Martina Iannacito \email{martina.iannacito@inria.fr}, 
Duccio Rocchini \email{duccio.rocchini@unibo.it}
}
\keyword{internal}
