% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShannonP.R
\name{ShannonP}
\alias{ShannonP}
\title{Parallelised Shannon's diversity index}
\usage{
ShannonP(rasterm, w, na.tolerance, debugging)
}
\arguments{
\item{rasterm}{Input data; a matrix of raster data values.}

\item{w}{Numeric; half of the side of the square moving window used in the calculation.}

\item{na.tolerance}{Numeric; a tolerance level (between 0.0 and 1.0) for NA values in the moving
window. If the proportion of NA values in a window exceeds this level, the result for that
window is set as NA; otherwise, the calculation excludes the NA values. The default is 0.0,
which allows no NA values.}

\item{debugging}{Boolean; if TRUE, additional debugging information is printed during the 
function's execution. This is helpful for debugging and is FALSE by default.}
}
\value{
A matrix or a list of matrices, each containing the Shannon diversity index values
calculated using a moving window approach.
}
\description{
This function computes Shannon's diversity index in a parallelized manner, allowing for
enhanced performance on suitable hardware. It is especially useful for large datasets.
}
\seealso{
\code{\link{Shannon}} for the non-parallelized version of the Shannon diversity index.
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, 
Martina Iannacito \email{martina.iannacito@inria.fr}, 
Duccio Rocchini \email{duccio.rocchini@unibo.it}
}
