% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShannonS.R
\name{ShannonS}
\alias{ShannonS}
\title{Sequential Shannon's diversity index}
\usage{
ShannonS(rasterm, w, na.tolerance, debugging)
}
\arguments{
\item{rasterm}{Input data; typically a matrix of raster data values.}

\item{w}{Numeric; half of the side of the square moving window. It determines the area over which each 
Shannon index value is calculated.}

\item{na.tolerance}{Numeric; a tolerance threshold (between 0.0 and 1.0) for NA values in the moving 
window. If the proportion of NA values in a window exceeds this threshold, the result for that 
window is set as NA. Otherwise, the calculation ignores the NA values. The default is 0.0, 
indicating no tolerance for NA values.}

\item{debugging}{Boolean; if TRUE, the function outputs additional information useful for debugging 
purposes. This parameter is FALSE by default, meaning that the extra information is not displayed 
during normal operation.}
}
\value{
A matrix or a list of matrices, each containing the Shannon diversity index values
calculated over the corresponding area of the input raster data.
}
\description{
This function computes Shannon's diversity index using a sequential method. It is designed for situations 
where parallel computation is not feasible or desired. The function applies a moving window approach to 
the provided raster data.
}
\seealso{
\code{\link{Shannon}} for the standard (non-sequential) version of the Shannon diversity index calculation.
}
\author{
Matteo Marcantonio \email{marcantoniomatteo@gmail.com}, 
Martina Iannacito \email{martina.iannacito@inria.fr}, 
Duccio Rocchini \email{duccio.rocchini@unibo.it}
}
