% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{validateInputs}
\alias{validateInputs}
\title{Validate Input Parameters for Diversity Index Calculation}
\usage{
validateInputs(x, window, alpha = 1, na.tolerance)
}
\arguments{
\item{x}{Raster object to be validated.}

\item{window}{Size of the moving window for calculations.}

\item{alpha}{Diversity index parameter, default is 1.}

\item{na.tolerance}{Proportion of acceptable NA values within the window (range: 0 to 1).}
}
\value{
None. Throws an error if any input is invalid.
}
\description{
Validates the input parameters for diversity index calculation functions. Checks for valid raster types, window sizes, alpha values, and NA tolerance levels.
}
