\docType{data}
\name{world}
\alias{world}
\title{Natural Earth world dataset}
\format{SpatialPolygonsDataFrame:
  \describe{
    \item{world}{SpatialPolygonsDataFrame of the world dissolved on continents.}
}}
\usage{
copNDVI
}
\description{
A SpatialPolygonsDataFrame (EPSG: 4326) of the continents.
}
\references{
\url{https://www.naturalearthdata.com/}
}
\keyword{datasets}