% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paRaoAreaS.R
\name{paRaoAreaS}
\alias{paRaoAreaS}
\title{Area-Based Sequential Parametric Rao's index of quadratic entropy (Q)}
\usage{
paRaoAreaS(rasterm, area, alpha, simplify)
}
\arguments{
\item{rasterm}{input raster data.}

\item{area}{input vector data.}

\item{alpha}{alpha value for order of diversity in Hill's Index.}

\item{simplify}{Rounding parameter.}
}
\value{
A vector as the input with Rao's index pasted additional columns.
}
\description{
Area-Based Sequential Parametric Rao's index of quadratic entropy (Q).
}
\seealso{
\code{\link{paRao}}
}
\author{
Marcantonio Matteo \email{marcantoniomatteo@gmail.com}, 
Duccio Rocchini \email{duccio.rocchini@unibo.it}, 
Michele Torresani \email{michele.torresani@unibo.it}
}
\keyword{internal}
